% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot-machine.R
\name{get_symbols}
\alias{get_symbols}
\title{Get random symbol from vector}
\usage{
get_symbols(...)
}
\value{
obj a vector of 3 element character
}
\description{
simulate the return of the slot machine value
also declare the probability of return of each symbol
}
\examples{
get_symbols()
}
