% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot-machine.R
\name{print.slotsMachine}
\alias{print.slotsMachine}
\title{S3 print method for class 'slotsMachine'}
\usage{
\method{print}{slotsMachine}(x, ...)
}
\description{
need to be name print.slotsMachine
method need to take the same argument as print. `args(print)`
}
