% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_dual_yaxis.R
\name{scale_factory}
\alias{scale_factory}
\title{function factory to assign scale function}
\usage{
scale_factory(X)
}
\arguments{
\item{X}{parameters for the factory function}
}
\value{
a function with saved params
}
\description{
function factory to assign scale function
}
