% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slot-machine.R
\name{score}
\alias{score}
\title{Score the prize from symbols}
\usage{
score(symbols)
}
\arguments{
\item{symbols}{a vector of 3 element}
}
\value{
`prize` money value
}
\description{
calculate prize based on multiple condition
}
