% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_misc}
\alias{theme_misc}
\title{Add custom ggplot theme}
\usage{
theme_misc()
}
\description{
This function will add a custom ggplot theme if add to a ggplot object.
It will have
* font "sans"="arial", size=6
* transparent panel background with a black border
* grid with grey color
}
\examples{
df <- data.frame(dose=c("D0.5", "D1", "D2"), len=c(4.2, 10, 29.5))
p <- ggplot(data=df, aes(x=dose, y=len)) + geom_bar(stat="identity")
p + theme_misc()
}
